Introduced base classes for MiKo_3222, MiKo_3223 and MiKo_3224
